:wrench: Fixed windows compilation (again).
